Fill out metadata comment to fix a bug in merge script.
We can't leave it empty since since "" is the same msgid as
the .po files header. So when we merge translations back into the
metadata.desktop, it uses the entire po header as the translation for
Comment[fr]="Plas...".

I need to find a fix as it should be possible to leave the comment
empty in a new widget until filled out.